Refactor features and defaults into module
